Fixed incorrect type usage of float (System.Single).

+++ CsLuaConverter/CsLuaConverter/CodeTreeLuaVisitor/Expression/Literal/NumericLiteralExpressionVisitor.cs
@@ -30,7 +30,7 @@
 
         private Type DetermineType()
         {
-            return this.value.Contains(".") ? typeof (float) : typeof (int);
+            return this.value.Contains(".") ? typeof (double) : typeof (int);
         }
     }
 }
